Restructuring. pRF overlap calculation adjusted for PacMan stimulus.

+++ 20161221_distcor_func/08_pRF_pre_and_post/utilities.py
@@ -44,7 +44,7 @@ def fncLoadNii(strPathIn):
     # Load nii file (this doesn't load the data into memory yet):
     objNii = nb.load(strPathIn)
     # Load data into array:
-    aryNii = objNii.get_data()
+    aryNii = objNii.get_data().astype(np.float32)
     # Get headers:
     objHdr = objNii.header
     # Get 'affine':
